pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "auth>=0.5.3",
    "bcrypt>=4.3.0",
    "cairosvg>=2.8.2",
    "fastapi>=0.116.0",
    "kaleido>=1.0.0",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.3.0",
    "plotly>=6.2.0",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.2",
    "scikit-learn>=1.7.0",
    "sendgrid>=6.12.4",
    "sqlalchemy>=2.0.41",
    "twilio>=9.6.3",
    "uvicorn[standard]>=0.35.0",
]
